Filter out invalid IP addresses.
When the Avahi daemon detects Discovery Controllers over mDNS, it
will report all IP address families (IPv4 and IPv6) being
advertized. However, the host may not be configured to support all
IP address families being advertized and should only try to connect
with the IP address family that is supported for a given interface.
This patch solves this issue.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>