Aligning code with upstream spec files.
I was looking at SuSE's and Red Hat's RPM spec file and just
wanted to align the sample spec file provided with nvme-stas
with the upstream projects. It can't be 100% aligned because
each distro has its own way of doing things, but at least
the sample spec file is now a bit similar.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>