stacd: allow disabling nvme-cli auto-connect udev rule.
nvme-cli installs a udev rule (70-nvmf-autoconnect.rules) that is
used to automatically connect to I/O Controllers (IOC) as a result
of receiving an AEN indicating a change of Discovery Log Page
Entries (DLPE) from a Discovery Controller.

This udev rule creates a race condition with stafd/stacd who also
react to AENs to automatically connect to IOCs. This is not a
critical problem since the connections are successful and stacd is
designed to handle potential race conditions. However, error
messages may get printed to the syslog even though everything is
working fine.

This patch allows disabling the udev rule to eliminate the race
condition. The default is to leave the udev rule active and allow
for the race condition to happen. If a user gets annoyed by the
false error messages, they can configure stacd.conf to disable the
udev rule.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>