staslib: calling wrong cback function on controller removal.
There are two cback functions that can be called when a controller
is removed. One of which is only meant to be called on process
shutdown when disconnecting from all controllers. The other cback
is meant to be called during normaln operation when we need to
disconnect from a single controller. We were calling the wrong
cback during normal disconnect, which led the process to think we
were shutting down.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>