test: Fix problem between fakefs and when configuration gets loaded
When running tests, we sometimes use fakefs to fake configuration
files such as /etc/nvme/hostnqn. It's important that the configuration
singletons such as conf.SysConf() do not get created before the fakefs
has been put in place otherwise the configuration will be wrong.

This was happening because "import stas" was actually initializing
the configuration before we had time to configure the fake fs. The
solution is to never implicitly intantiate singletons (such as
conf.SysConf()) during an "import", but rather explicitly instantiate
with a function call.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>