stacd: Add support for new src_addr attribute in sysfs
A patch was submitted to the Linux kernel to expose the actual
source address of a connected TCP socket. This will allow stacd to
infer the interface associated with a connection that was made by
relying on the routing table to selct the interface. This, in turn,
will provide better interop with nvme-cli.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>