stafd: Handle unresponsive DCs in _config_ctrls_finish()
This is so that handling of all configuration changes, be it to
add or remove controllers, will be handled in one central place,
which is function _config_ctrls_finish().

Also, since _config_ctrls_finish() is a callback function, it needs
to check whether the service is still alive before proceeding with
creation/removal of controller objects. Previously, there was the
chance of a race condition if _config_ctrls_finish() got called
after a long delay and if during that delay the service got stopped
(e.g. systemctl stop stafd). To prevent that, the first line of
_config_ctrls_finish() now checks that the service is still alive
before proceeding. The same logic was added to all callback
functions to prevent operating on a service that is being stopped.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>