staslib: fix RoCE support
NVMe TP 8009 mentions using UDP for RoCE. However, nvme-stas currently
registers for the TCP discovery service alone, and not UDP. So while
avahi can see the RoCE service on the target, nvme-stas doesn't ask to
be alerted to them. So update avahi to handle the same. Also, nvme-stas
should convert p=roce in TXT to rdma when creating discovery
controllers. And finally, ensure host-iface is avoided for RDMA
controllers since it is not permitted in the kernel.

Signed-off-by: Martin George <marting@netapp.com>