stafd|ctrl: Do not recreate ctrl object on nvme removed events
On kernel events indicating that an nvme device associated with
a discovery controller was removed by the kernel, stafd should not
delete the old ctrl object and create a new one to kick-start the
reconnect process. This causes the cached DLPEs to be lost, which
may result in stacd disconnecting from I/O controllers. Instead,
simply restart the "try to connect" timer, while stafd is trying
to reconnect, the DLPEs will be preserved, which will allow stacd
to maintain existing I/O controller connections.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>