trid: Take Host NQN into account
When booting from SAN (BFS), the Host NQN specified in the NBFT
files is often different from the Host NQN specified in
/etc/nvme/hostnqn. Therefore, connections made after the system
is fully booted will be made with a different Host NQN than the
connections made during boot. nvme-stas did not consider the Host
NQN when scanning for existing connections, which means that it
would reuse a BFS connection when in fact it needed to create a
new one using a different Host NQN.

Note that the same issue is seen in nvme-cli and libnvme as well.
These will need to be fixed separately.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>