avahi: Add retry mechanism on TCP checker code
We found that SFSS advertises (mDNS) its services before it is even
listening for connections. If a host responds too quickly to Avahi
detection of the CDC and tries to connect before the CDC is actually
listening, we will get the connection refused error.

SFSS should only advertise its service with mDNS when it is ready
to receive connections. However, nvme-stas will now have a retry
mechanism in case SFSS cannot fix this issue.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>