Skip mac2iface test for esoteric interfaces
mac2iface takes a MAC address as argument and returns the corresponding
interface (if any).

The mac2iface tests will however invoke mac2iface with invalid MAC
addresses when esoteric network interfaces are present on the system. As
an example, armhf autopkgtest runners in Ubuntu have gre interfaces
configured. gre interfaces do not store a MAC address in the "address"
field:

 $ ip link add type gre
 $ ip -j addr show dev gre0 | jq '.[]["address"]'
    "0.0.0.0"

We now ensure that the test-suite calls mac2iface with only valid MAC
addresses. We also skip all the 00:00:00:00:00:00 addresses because they
are invalid and can be visible on multiple interfaces (not just the
loopback interface).

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>