Skip mac2iface test when multiple interfaces have the same MAC
Sometimes, the same MAC address is assigned to more than one interface
on the system (this is true for VLAN interfaces for instance). This
confuses mac2iface, which returns only the first match.

This scenario is possibly more of a nvme-stas bug than a test-suite bug,
but for now we will just skip the interfaces that have a duplicate MAC
address.

 $ ip link add link eth0 name eth0.222 type vlan id 222
 $ ip -j addr show dev eth0 | jq '.[]["address"]'
    "00:16:3e:36:b5:0e"
 $ ip -j addr show dev eth0.222 | jq '.[]["address"]'
    "00:16:3e:36:b5:0e"

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>