Build: Fix Github actions
The environment used to run GitHub actions has been updated
upstream, which breaks the build. This is mainly due to the use
of pip. That is, the use of upstream pip and not the one provided
by Debian through apt-get install python3-pip.

This updates all the build scripts to use the standard Debian
packages (if available) and not rely on "pip install".

Signed-off-by: Martin Belanger <martin.belanger@dell.com>